data types changed

+++ Calculator/Calculator/main.m
@@ -10,24 +10,24 @@
 @interface Calculator: NSObject
 @property int n1, n2;
 
--(double) sum: (double) n1 over: (double) n2;
--(double) substract: (double) n1 over: (double) n2;
--(double) multiply: (double) n1 over: (double) n2;
+-(int) sum: (double) n1 over: (double) n2;
+-(float) substract: (double) n1 over: (double) n2;
+-(long) multiply: (double) n1 over: (double) n2;
 -(double) divide: (double) n1 over: (double) n2;
 
 @end
 
 @implementation Calculator
 
--(double) sum:(double) n1 over: (double) n2 {
+-(int) sum:(double) n1 over: (double) n2 {
         double sum = n1 + n2;
         return sum;
     }
--(double) substract:(double)n1 over:(double)n2 {
+-(float) substract:(double)n1 over:(double)n2 {
     double substract = n1 - n2;
     return substract;
 }
--(double) multiply:(double)n1 over:(double)n2 {
+-(long) multiply:(double)n1 over:(double)n2 {
     double multiply = n1 * n2;
     return multiply;
 }
@@ -42,9 +42,9 @@ int main(int argc, char* argv[]) {
     @autoreleasepool {
         Calculator *calculator = [[Calculator alloc] init];
 
-        NSLog(@"\nSum is %f", [calculator sum:(16.8) over: (8.2)]);
+        NSLog(@"\nSum is %i", [calculator sum:(16.8) over: (8.2)]);
         NSLog(@"\nThe difference is %f", [calculator substract:(16.8) over: (8.2)]);
-        NSLog(@"\nProduct is %f", [calculator multiply:(16.8) over: (8.2)]);
+        NSLog(@"\nProduct is %ld", [calculator multiply:(16.8) over: (8.2)]);
         NSLog(@"\nDivision is %f", [calculator divide:(16.8) over: (8.2)]);
         
     }
